Adding tuple visibility function to storage AM
Tuple visibility functions are now part of the
heap storage AM routine. The visibilty execution
procedure is changed accoridngly.

The snapshot satifies function is changed to an
enum to represent what type of snapshot is it
and this enum value is used to call the corresponding
visibilty function from the storage AM when the
visibilty of the tuple is required.

The common code is that is part of both server
and pluggable storages is now moved into storage_common.c
and storage_common.h files.