Adding tuple visibility functions to table AM
Tuple visibility functions are now part of the
heap table AM routine. The visibilty execution
procedure is changed accoridngly.

The snapshot satifies function is changed to an
enum to represent what type of snapshot is it
and this enum value is used to call the corresponding
visibilty function from the storage AM when the
visibilty of the tuple is required.

The common code is that is part of both server
and pluggable table access methods is now moved
into tableam_common.h file.