slot hooks are added to table AM
The tuple is removed as part of the slot and added
an void pointer to store the tuple data that can
understand only by the table AM routine.

The slot utility functions are reorganized to use
two table AM routines to satify the current
functionality.

Currently the slot supports minimum tuple also.