Server recovery mode handling
in_recovery GUC_REPORT is added to update the clients when the
server is recovery mode, this is useful for the client connections
to connect to a standby server with a faster check instead of
executing a command.

New SIGUSR1 handling interrupt is added to support reporting
of recovery mode exit to all backends and their respective
clients.

Some parts of the code is taken from earlier development by
Elvis Pranskevichus and Tsunakawa Takayuki.